Whew that's a lot.
* Swapped a lot of int primary keys for long, mostly for uniformity partially for paranoia.
* Added the Group model, not wired into the form or repository yet
* Views now have a more abstract ActivityOwner concept which could be a client or a group.
* That ActivityOwner concept exposed a type bug in ActivityEditor - it was getting summary clients not full on clients.

As a note, I originally tried to do this with a ClientActivity and GroupActivity table with all the common fields in a base Activity class. That way both would have a non-null foreign key and avoid the ambiguous cascade delete problem. EF Core 6 HATES this and kept insisting on TPH modeling. EF Core 7 introduces TPC but I don't want to leave the LTS version. With only two owners (group / client) this is probably going to be fine.